requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-WTF==1.1.1
mysql-connector-python==8.1.0
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
requests==2.31.0
selenium==4.15.2
pytest==7.4.3
pytest-cov==4.1.0
webdriver-manager==4.0.1
coverage==7.4.1
sshtunnel==0.4.0 